requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0

# Kafka
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP
httpx==0.26.0
requests==2.31.0

# Data & ML
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
prophet==1.1.5
scikit-learn==1.4.0
joblib==1.3.2

# Stock Data
yfinance==0.2.36
kiteconnect==5.0.1

# Utils
python-dotenv==1.0.0
pytz==2024.1
websocket-client==1.7.0
websockets==12.0